feat(app): minor billing improvements
