+++ apps/app/src/lib/date.ts
@@ -1,11 +1,12 @@
-import dayjs, { extend } from "dayjs";
+import dayjs, { type Dayjs, extend } from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
 
 extend(relativeTime);
 
-export const timeFromNow = (date: string): string => dayjs(date).fromNow();
-export const monthDay = (date: string): string => dayjs(date).format("MMM D");
-export const monthDayYear = (date: string): string => dayjs(date).format("MMM D, YYYY");
+export const timeFromNow = (date: string | Dayjs | Date): string => dayjs(date).fromNow();
+export const monthDay = (date: string | Dayjs | Date): string => dayjs(date).format("MMM D");
+export const monthDayYear = (date: string | Dayjs | Date): string =>
+  dayjs(date).format("MMM D, YYYY");
 
 // TODO: replace with parseIsoDate
 export const isIsoDate = (date: string): boolean => {
