+++ packages/shared/src/index.ts
@@ -1,4 +1,3 @@
 export * from "./billing";
 export * from "./format-number";
 export * from "./links";
-export * from "./pricing-consts";
